+++ packages/related-articles/related-articles.stories.js
@@ -2,7 +2,7 @@ import { storiesOf } from "@storybook/react-native";
 import React from "react";
 import { ScrollView } from "react-native";
 import { decorateAction } from "@storybook/addon-actions";
-import { storybookReporter } from "@times-components/tealium";
+import storybookReporter from "@times-components/tealium-utils";
 import RelatedArticles from "./src/related-articles";
 
 import standard1RelatedArticleFixture from "./fixtures/standard/1-article.json";