+++ packages/tealium-utils/src/index.js
@@ -1,0 +1,3 @@
+import storybookReporter from "./storybook";
+
+export default storybookReporter;