+++ packages/tealium/src/index.js
@@ -1,4 +1,3 @@
 import createTealiumReport from "./tealium";
 
-export { default as storybookReporter } from "./storybook";
 export default createTealiumReport;