+++ packages/topics/__tests__/shared.js
@@ -8,7 +8,7 @@ import topicData from "../fixtures/topics";
 
 module.exports = () => {
   beforeEach(() => {
-    mockDate.set("1/1/2018");
+    mockDate.set(1514764800000, 0);
   });
 
   afterEach(() => {