+++ packages/utils/src/index.js
@@ -1,4 +1,5 @@
 export * from "./faketime";
 export * from "./graphql";
 export * from "./strings";
+export { default as addMissingProtocol } from "./add-missing-protocol";
 export { default as schema } from "./schema.json";