+++ packages/video-label/__tests__/shared.js
@@ -4,7 +4,7 @@ import renderer from "react-test-renderer";
 import VideoLabel from "../src/video-label";
 
 export default () => {
-  it("renders VideoLabel", () => {
+  it("renders VideoLabel with an explicit title", () => {
     const tree = renderer
       .create(<VideoLabel title="swimming" color="#008347" />)
       .toJSON();
@@ -12,10 +12,8 @@ export default () => {
     expect(tree).toMatchSnapshot();
   });
 
-  it("does not render title if title equals Video", () => {
-    const tree = renderer
-      .create(<VideoLabel title="Video" color="#008347" />)
-      .toJSON();
+  it("renders VideoLabel without an explicit title", () => {
+    const tree = renderer.create(<VideoLabel color="#008347" />).toJSON();
 
     expect(tree).toMatchSnapshot();
   });