+++ packages/video-label/video-label.stories.js
@@ -1,7 +1,7 @@
 import "react-native";
 import React from "react";
 import { storiesOf } from "@storybook/react-native";
-import { select } from "@storybook/addon-knobs/react";
+import { select } from "@times-components/storybook";
 import invert from "lodash.invert";
 import { colours } from "@times-components/styleguide";
 import VideoLabel from "./src/video-label";