+++ packages/author-head/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.23.1"></a>
+## [0.23.1](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.23.0...@times-components/author-head@0.23.1) (2018-04-12)
+
+
+### Bug Fixes
+
+* do not publish on npm settings files ([#848](https://github.com/newsuk/times-components/issues/848)) ([a1df7f3](https://github.com/newsuk/times-components/commit/a1df7f3))
+
+
+
+
+<a name="0.23.0"></a>
+# [0.23.0](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.24...@times-components/author-head@0.23.0) (2018-04-12)
+
+
+### Features
+
+* add tealium-utils package ([#846](https://github.com/newsuk/times-components/issues/846)) ([b5da5fa](https://github.com/newsuk/times-components/commit/b5da5fa))
+
+
+
+
+<a name="0.22.24"></a>
+## [0.22.24](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.23...@times-components/author-head@0.22.24) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
 <a name="0.22.23"></a>
 ## [0.22.23](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.22.22...@times-components/author-head@0.22.23) (2018-04-11)