+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.16"></a>
+## [2.1.16](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.15...@times-components/brightcove-video@2.1.16) (2018-04-12)
+
+
+### Bug Fixes
+
+* do not publish on npm settings files ([#848](https://github.com/newsuk/times-components/issues/848)) ([a1df7f3](https://github.com/newsuk/times-components/commit/a1df7f3))
+
+
+
+
+<a name="2.1.15"></a>
+## [2.1.15](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.14...@times-components/brightcove-video@2.1.15) (2018-04-12)
+
+
+### Bug Fixes
+
+* add protocol to splash REPLAT-1677 ([#837](https://github.com/newsuk/times-components/issues/837)) ([109f0f5](https://github.com/newsuk/times-components/commit/109f0f5))
+
+
+
+
 <a name="2.1.14"></a>
 ## [2.1.14](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.1.13...@times-components/brightcove-video@2.1.14) (2018-04-11)