+++ packages/depend/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.13"></a>
+## [0.1.13](https://github.com/newsuk/times-components/compare/@times-components/depend@0.1.12...@times-components/depend@0.1.13) (2018-04-12)
+
+
+### Bug Fixes
+
+* do not publish on npm settings files ([#848](https://github.com/newsuk/times-components/issues/848)) ([a1df7f3](https://github.com/newsuk/times-components/commit/a1df7f3))
+
+
+
+
 <a name="0.1.12"></a>
 ## [0.1.12](https://github.com/newsuk/times-components/compare/@times-components/depend@0.1.11...@times-components/depend@0.1.12) (2018-04-11)