+++ packages/jest-configurator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.2"></a>
+## [0.5.2](https://github.com/newsuk/times-components/compare/@times-components/jest-configurator@0.5.1...@times-components/jest-configurator@0.5.2) (2018-04-12)
+
+
+### Bug Fixes
+
+* do not publish on npm settings files ([#848](https://github.com/newsuk/times-components/issues/848)) ([a1df7f3](https://github.com/newsuk/times-components/commit/a1df7f3))
+
+
+
+
 <a name="0.5.1"></a>
 ## [0.5.1](https://github.com/newsuk/times-components/compare/@times-components/jest-configurator@0.5.0...@times-components/jest-configurator@0.5.1) (2018-04-11)