+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.1"></a>
+## [0.1.1](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.1.0...@times-components/related-articles@0.1.1) (2018-04-12)
+
+
+### Bug Fixes
+
+* do not publish on npm settings files ([#848](https://github.com/newsuk/times-components/issues/848)) ([a1df7f3](https://github.com/newsuk/times-components/commit/a1df7f3))
+
+
+
+
+<a name="0.1.0"></a>
+# [0.1.0](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.17...@times-components/related-articles@0.1.0) (2018-04-12)
+
+
+### Features
+
+* add tealium-utils package ([#846](https://github.com/newsuk/times-components/issues/846)) ([b5da5fa](https://github.com/newsuk/times-components/commit/b5da5fa))
+
+
+
+
+<a name="0.0.17"></a>
+## [0.0.17](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.16...@times-components/related-articles@0.0.17) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.16"></a>
+## [0.0.16](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.15...@times-components/related-articles@0.0.16) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.15"></a>
+## [0.0.15](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.14...@times-components/related-articles@0.0.15) (2018-04-11)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
 <a name="0.0.14"></a>
 ## [0.0.14](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.13...@times-components/related-articles@0.0.14) (2018-04-11)