+++ packages/tealium/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/newsuk/times-components/compare/@times-components/tealium@0.3.0...@times-components/tealium@0.3.1) (2018-04-12)
+
+
+### Bug Fixes
+
+* do not publish on npm settings files ([#848](https://github.com/newsuk/times-components/issues/848)) ([a1df7f3](https://github.com/newsuk/times-components/commit/a1df7f3))
+
+
+
+
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/newsuk/times-components/compare/@times-components/tealium@0.2.22...@times-components/tealium@0.3.0) (2018-04-12)
+
+
+### Features
+
+* add tealium-utils package ([#846](https://github.com/newsuk/times-components/issues/846)) ([b5da5fa](https://github.com/newsuk/times-components/commit/b5da5fa))
+
+
+
+
+<a name="0.2.22"></a>
+## [0.2.22](https://github.com/newsuk/times-components/compare/@times-components/tealium@0.2.21...@times-components/tealium@0.2.22) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium
+
 <a name="0.2.21"></a>
 ## [0.2.21](https://github.com/newsuk/times-components/compare/@times-components/tealium@0.2.20...@times-components/tealium@0.2.21) (2018-04-11)